--- src/main/java/com/tavemakers/surf/domain/login/kakao/config/RestClientConfig.java
@@ -1,19 +1,0 @@
-package com.tavemakers.surf.domain.login.kakao.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import java.time.Duration;
-
-@Configuration
-public class RestClientConfig {
-
-    @Bean
-    public RestTemplate kakaoRestTemplate(RestTemplateBuilder builder) {
-        return builder
-                .setConnectTimeout(Duration.ofSeconds(2))
-                .setReadTimeout(Duration.ofSeconds(5))
-                .build();
-    }
-}
