--- src/main/java/com/tavemakers/surf/domain/login/kakao/dto/KakaoUserInfoDTO.java
@@ -1,31 +1,0 @@
-package com.tavemakers.surf.domain.login.kakao.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-/**
- * Kakao 사용자 정보 DTO
- * - 카카오 API /v2/user/me 응답 JSON을 매핑
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserInfoDTO(
-        Long id,
-        KakaoAccount kakaoAccount
-) {
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-    public record KakaoAccount(
-            String email,
-            Profile profile
-    ) {}
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-    public record Profile(
-            String nickname,
-            String profileImageUrl,
-            String thumbnailImageUrl
-    ) {}
-}
